% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_matrix.R
\name{adj_matrix}
\alias{adj_matrix}
\title{Create Adjacency Matrix}
\usage{
adj_matrix(seu, k = 7)
}
\arguments{
\item{seu}{Seurat-Object}

\item{k}{distance to neighbours}
}
\value{
The Adjacency Matrix
}
\description{
Create Adjacency Matrix
}
\keyword{internal}
