% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_nn_spatial_smooth.R
\name{alpha_nn_spatial_smooth}
\alias{alpha_nn_spatial_smooth}
\title{alpha_nn_spatial_smooth}
\usage{
alpha_nn_spatial_smooth(
  seu,
  genes = "Datasets - Ikarus - Gene_lists.csv",
  assay = "Spatial",
  alpha = 0.4,
  a = 0.8,
  graph = "nn"
)
}
\arguments{
\item{seu}{a Seurat-Object}

\item{genes}{csv-file indicating which genes to use (optional)}

\item{assay}{which assay to use (optional)}

\item{alpha}{for linear-combination alpha*spatial + (1-alpha)*expression}

\item{a}{alpha for smoothing (optional)}

\item{graph}{which neighbour-graph to use, "nn" or "snn"}
}
\value{
vector with smoothed scores
}
\description{
alpha_nn_spatial_smooth
}
\examples{
alpha_nn_spatial_smooth(se, assay= "RNA")
}
