% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaCalc.R
\name{gseaCalc}
\alias{gseaCalc}
\title{calculate gsea-score}
\usage{
gseaCalc(se, genes = "Datasets - Ikarus - Gene_lists.csv", assay = "Spatial")
}
\arguments{
\item{se}{Seurat object or VoltRon object}

\item{genes}{csv-file indicating which genes to use (optional)}

\item{assay}{which assay to use (optional)}
}
\value{
Seurat/VoltRon object with calculated scores in meta.data
}
\description{
calculate gsea-score
}
\examples{
gseaCalc(seu, assay="RNA")
}
