% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_smooth.R
\name{nn_smooth}
\alias{nn_smooth}
\title{perform nn smoothing}
\usage{
nn_smooth(
  seu,
  genes = "V1_Breast_Cancer_Block_A_Section_2_spatial/Datasets - Ikarus - Gene_lists.csv",
  assay = "Spatial",
  a = 0.8,
  graph = "nn"
)
}
\arguments{
\item{seu}{Seurat object}

\item{genes}{csv-file indicating which genes to use (optional)}

\item{assay}{which assay to use (optional)}

\item{a}{alpha for smoothing (optional)}

\item{graph}{which neighbour-graph to use, "nn" or "snn"}
}
\value{
vector with smoothed scores
}
\description{
perform nn smoothing
}
