% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_spatial_smooth.R
\name{nn_spatial_smooth}
\alias{nn_spatial_smooth}
\title{nn_spatial_smooth}
\usage{
nn_spatial_smooth(
  se,
  genes = "Datasets - Ikarus - Gene_lists.csv",
  assay = "Spatial",
  a1 = 0.8,
  a2 = 0.8,
  graph = "nn",
  k = 7
)
}
\arguments{
\item{genes}{csv-file indicating which genes to use (optional)}

\item{assay}{which assay to use (optional)}

\item{a1}{alpha for nn/snn smoothing (optional)}

\item{a2}{alpha for spatial smoothing (optional)}

\item{graph}{which neighbour-graph to use, "nn" or "snn"}

\item{k}{distance of cell to neighbors}

\item{seu}{Seurat-Object or VoltRon object}
}
\value{
vector with smoothed scores
}
\description{
nn_spatial_smooth
}
\examples{
nn_spatial_smooth(seu, a1 = 0.5, a2 = 0.8, graph = "nn")
}
