% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quant.R
\name{plot_quant}
\alias{plot_quant}
\title{plot_quant}
\usage{
plot_quant(gsea_score, coordinates, truth)
}
\arguments{
\item{gsea_score}{vector of calculated score for each cell}

\item{coordinates}{tissue coordinates}

\item{truth}{vector for all cells with 1 for tumor and 0 for normal cells}
}
\value{
plot of truth of prediction thresholded on ROC-curve
}
\description{
plot_quant
}
\examples{
plot_quant(gsea, coords, truth)
}
