% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_quant.R
\name{roc_quant}
\alias{roc_quant}
\title{roc_quant}
\usage{
roc_quant(truth, scores)
}
\arguments{
\item{truth}{vector of truth, 0 indicating normal cells, 1 cells with condition}

\item{scores}{vector of calculated scores}
}
\value{
return best threshold and accuracy
}
\description{
roc_quant
}
\examples{
roc_quant(c(0,1,1), c(-0.03, 0.6, 0.87))
}
