% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_smooth.R
\name{spatial_smooth}
\alias{spatial_smooth}
\title{perform spatial smoothing}
\usage{
spatial_smooth(
  seu,
  genes = "V1_Breast_Cancer_Block_A_Section_2_spatial/Datasets - Ikarus - Gene_lists.csv",
  assay = "Spatial",
  a = 0.8
)
}
\arguments{
\item{seu}{Seurat object}

\item{genes}{csv-file indicating which genes to use (optional)}

\item{assay}{which assay to use (optional)}

\item{a}{alpha for smoothing (optional)}
}
\value{
vector with smoothed scores
}
\description{
perform spatial smoothing
}
