% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_smooth.R
\name{union_smooth}
\alias{union_smooth}
\title{union smoothing}
\usage{
union_smooth(
  seu,
  genes = "Datasets - Ikarus - Gene_lists.csv",
  assay = "Spatial",
  a = 0.8,
  graph = "nn",
  k = 7
)
}
\arguments{
\item{seu}{Seurat-Object or VoltRon object}

\item{genes}{csv-file indicating which genes to use (optional)}

\item{assay}{which assay to use (optional)}

\item{a}{alpha for smoothing (optional)}

\item{graph}{which neighbour-graph to use, "nn" or "snn"}

\item{k}{distance of cell to neighbors}
}
\value{
vector with smoothed scores
}
\description{
union smoothing
}
\examples{
union_smooth(seu, a = 0.4)
}
