% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{volt}
\alias{volt}
\title{Human breast-cancer-tissue}
\format{
A VoltRon-Object
}
\source{
a paper by Anderson et al. (2021). The processed count matrices, HE-images and metadata with spot-annotation can be found here: https://zenodo.org/records/4751624. The sample used is B1
data(volt)
}
\usage{
volt
}
\description{
A VoltRon-Object containing spatial and gene-expression data of human breats-tissue with cancer
}
\keyword{datasets}
