% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneset_ikarus}
\alias{geneset_ikarus}
\title{List of human genes}
\format{
A list of genes with with Type "Tumor" or "Normal"
}
\source{
?
}
\usage{
geneset_ikarus
}
\description{
Genes where high expression is 
either associated with normal cells or cancer cells
obtained using Ikarus
https://github.com/BIMSBbioinfo/ikarus
}
\keyword{datasets}
