% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_smooth.R
\name{inter_smooth}
\alias{inter_smooth}
\title{intersection smoothing}
\usage{
inter_smooth(
  seu,
  genes = "Datasets - Ikarus - Gene_lists.csv",
  assay = "Spatial",
  a = 0.8,
  graph = "nn"
)
}
\arguments{
\item{seu}{Seurat-Object}

\item{genes}{csv-file indicating which genes to use (optional)}

\item{assay}{which assay to use (optional)}

\item{a}{alpha for smoothing (optional)}

\item{graph}{which neighbour-graph to use, "nn" or "snn"}
}
\value{
vector with smoothed scores
}
\description{
intersection smoothing
}
\examples{
union_smooth(seu, a = 0.4)
}
