% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quant.R
\name{plot_quant}
\alias{plot_quant}
\title{plot_quant}
\usage{
plot_quant(gsea_score, coordinates, truth, threshold)
}
\arguments{
\item{gsea_score}{calculated score for each cell}

\item{coordinates}{of cell}

\item{truth}{vector for all cells with 1 for tumor and 0 for normal cells}

\item{threshold}{which percentile to take as threshold (1-10)}
}
\value{
plot of truth of prediction
}
\description{
plot_quant
}
\examples{
plot_quant(gsea, coords, truth, 5)
}
