% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quant.R
\name{plot_quant}
\alias{plot_quant}
\title{plot_quant}
\usage{
plot_quant(gsea_score, coordinates, truth, threshold)
}
\arguments{
\item{coordinates}{}

\item{threshold}{which percentile to take as threshold (1-10)}
}
\value{
plot of truth of prediction
}
\description{
plot_quant
}
