% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seu}
\alias{seu}
\title{Human breast-cancer-tissue}
\format{
A Seurat-Object
}
\source{
from 10xGenomics support page, Human Breast Cancer (Block A Section 1) https://www.10xgenomics.com/datasets/
data(seu)
}
\usage{
seu
}
\description{
A Seurat-Object containing spatial and gene-expression data of human breats-tissue with cancer
}
\keyword{datasets}
